+++ core/triggers/interface_90_modSubtotal_subtotaltrigger.class.php
@@ -180,7 +180,7 @@ class Interfacesubtotaltrigger extends DolibarrTriggers
 		/* Refer to issue #379 */
 		if($action == 'LINEBILL_INSERT'){
 			static $TInvoices = array();
-			if ($TInvoices[$object->fk_facture] === null) {
+			if (!array_key_exists($object->fk_facture, $TInvoices) || (array_key_exists($object->fk_facture, $TInvoices) && $TInvoices[$object->fk_facture] === null)) {
 				$staticInvoice = new Facture($this->db);
 				if ($staticInvoice->fetch($object->fk_facture) < 0){
 					$object->error = $staticInvoice->error;
@@ -304,7 +304,7 @@ class Interfacesubtotaltrigger extends DolibarrTriggers
             }
 		    else
             {
-			    $subtotal_add_title_bloc_from_orderstoinvoice = (GETPOST('subtotal_add_title_bloc_from_orderstoinvoice', 'none') || GETPOST('createbills_onebythird', 'int'));
+			    $subtotal_add_title_bloc_from_orderstoinvoice = (GETPOST('subtotal_add_title_bloc_from_orderstoinvoice', 'none') && GETPOST('createbills_onebythird', 'int'));
 			    if (!empty($subtotal_add_title_bloc_from_orderstoinvoice))
 			    {
 				    global $subtotal_current_rang, $subtotal_bloc_previous_fk_commande, $subtotal_bloc_already_add_title, $subtotal_bloc_already_add_st;